Refactorización de código

+++ src/thymeleaf/src/main/java/com/optativa/thymeleaf/repositorio/UsuarioRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.optativa.thymeleaf.entidad.Usuario;
 
-public interface UsuarioRepositorio extends JpaRepository<Usuario, Integer>{
+public interface UsuarioRepositorio extends JpaRepository<Usuario, Long>{
 	Usuario findByNombre(String username);
 }
